Turned onClickAnalytics into hook (to check if cookie was accepted before attempting to log)
