+++ public_website/src/ui/components/header/AccountItem.tsx
@@ -3,12 +3,14 @@ import styled, { css } from 'styled-components'
 
 import { ArrowRight } from '../icons/ArrowRight'
 import { OutlinedText } from '../OutlinedText'
-import { onClickAnalytics } from '@/lib/analytics/helpers'
+import { useOnClickAnalytics } from '@/hooks/useOnClickAnalytics'
 import { AccountItemDropdownProps } from '@/types/props'
 import { Link } from '@/ui/components/Link'
 
 export function AccountItem(props: AccountItemDropdownProps) {
   const { color, url, emoji, label, eventName, eventOrigin } = props
+  const { onClickAnalytics } = useOnClickAnalytics()
+
   return (
     <StyledAccountItem>
       <Link
